Address PR review feedback from tskarhed
Implements the following changes based on code review:

1. **Optimize pre-fetching**: Only pre-fetch the first scope's defaultPath
   instead of all scopes, since only the first is used for expansion.
   Also skip nodes already in cache to avoid unnecessary API calls.

2. **Extract parent from defaultPath**: When applying scopes, prefer
   extracting the parent node from defaultPath (second-to-last node)
   rather than from parentNodeId or scope node's parent. This is more
   reliable and consistent with the defaultPath-first approach.

3. **Update ScopesInput tooltip**: Modified getScopesPath() to use
   defaultPath from scope metadata when available, falling back to
   walking the node tree. This ensures the tooltip reflects the
   correct path hierarchy.

4. **Consolidate test files**: Merged ScopesSelectorService.defaultPath.test.ts
   and ScopesSelectorService.pathHelpers.test.ts into the main
   ScopesSelectorService.test.ts file as separate describe blocks.
   This improves test organization and maintainability.

All 96 tests in ScopesSelectorService.test.ts pass.
All 26 tests in ScopesApiClient.test.ts pass.
