+++ public/app/features/scopes/selector/ScopesSelectorService.test.ts
@@ -77,7 +77,14 @@ describe('ScopesSelectorService', () => {
       }),
       fetchDashboards: jest.fn().mockResolvedValue([]),
       fetchScopeNavigations: jest.fn().mockResolvedValue([]),
-      fetchScopeNode: jest.fn().mockResolvedValue(mockNode),
+      fetchScopeNode: jest.fn().mockImplementation((id: string) => {
+        // Return undefined for empty string (root node)
+        if (id === '') {
+          return Promise.resolve(undefined);
+        }
+        return Promise.resolve(mockNode);
+      }),
+      fetchMultipleScopeNodes: jest.fn().mockResolvedValue([]),
     } as unknown as jest.Mocked<ScopesApiClient>;
 
     dashboardsService = {
@@ -1234,6 +1241,1040 @@ describe('ScopesSelectorService', () => {
 
       // Should fall back to scope navigation since scope node is undefined
       expect(locationService.push).toHaveBeenCalledWith('/d/dashboard1');
+    });
+  });
+
+  // Mock data for defaultPath and path helper tests
+  const regionNode: ScopeNode = {
+    metadata: { name: 'region-us-west' },
+    spec: {
+      nodeType: 'container',
+      title: 'US West',
+      parentName: '',
+      linkType: undefined,
+      linkId: undefined,
+    },
+  };
+
+  const countryNode: ScopeNode = {
+    metadata: { name: 'country-usa' },
+    spec: {
+      nodeType: 'container',
+      title: 'USA',
+      parentName: 'region-us-west',
+      linkType: undefined,
+      linkId: undefined,
+    },
+  };
+
+  const cityNode: ScopeNode = {
+    metadata: { name: 'city-seattle' },
+    spec: {
+      nodeType: 'container',
+      title: 'Seattle',
+      parentName: 'country-usa',
+      linkType: undefined,
+      linkId: undefined,
+    },
+  };
+
+  const datacenterNode: ScopeNode = {
+    metadata: { name: 'datacenter-sea-1' },
+    spec: {
+      nodeType: 'leaf',
+      title: 'SEA-1',
+      parentName: 'city-seattle',
+      linkType: 'scope',
+      linkId: 'scope-sea-1',
+    },
+  };
+
+  const scopeWithDefaultPath: Scope = {
+    metadata: { name: 'scope-sea-1' },
+    spec: {
+      title: 'Seattle Datacenter 1',
+      defaultPath: ['region-us-west', 'country-usa', 'city-seattle', 'datacenter-sea-1'],
+      filters: [],
+    },
+  };
+
+  const scopeWithoutDefaultPath: Scope = {
+    metadata: { name: 'scope-no-path' },
+    spec: {
+      title: 'No Path Scope',
+      filters: [],
+    },
+  };
+
+  const parentNode: ScopeNode = {
+    metadata: { name: 'parent' },
+    spec: {
+      nodeType: 'container',
+      title: 'Parent',
+      parentName: '',
+      linkType: undefined,
+      linkId: undefined,
+    },
+  };
+
+  const childNode: ScopeNode = {
+    metadata: { name: 'child' },
+    spec: {
+      nodeType: 'leaf',
+      title: 'Child',
+      parentName: 'parent',
+      linkType: 'scope',
+      linkId: 'test-scope',
+    },
+  };
+
+  const grandchildNode: ScopeNode = {
+    metadata: { name: 'grandchild' },
+    spec: {
+      nodeType: 'leaf',
+      title: 'Grandchild',
+      parentName: 'child',
+      linkType: 'scope',
+      linkId: 'test-scope-2',
+    },
+  };
+
+  /* eslint-disable @typescript-eslint/no-explicit-any */
+  // Tests for defaultPath functionality
+  // Note: Tests access protected updateState method via (service as any) casting to set up test state
+  describe('getScopeNodes', () => {
+    it('should return cached nodes when available', async () => {
+      // Pre-populate cache
+      (service as any).updateState({
+        nodes: {
+          'region-us-west': regionNode,
+          'country-usa': countryNode,
+        },
+      });
+
+      const result = await service.getScopeNodes(['region-us-west', 'country-usa']);
+
+      expect(result).toEqual([regionNode, countryNode]);
+      expect(apiClient.fetchMultipleScopeNodes).not.toHaveBeenCalled();
+    });
+
+    it('should fetch only non-cached nodes', async () => {
+      // Pre-populate cache with one node
+      (service as any).updateState({
+        nodes: {
+          'region-us-west': regionNode,
+        },
+      });
+
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([countryNode]);
+
+      const result = await service.getScopeNodes(['region-us-west', 'country-usa']);
+
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith(['country-usa']);
+      expect(result).toEqual([regionNode, countryNode]);
+    });
+
+    it('should maintain order of requested nodes', async () => {
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([cityNode, countryNode, regionNode]);
+
+      const result = await service.getScopeNodes(['region-us-west', 'country-usa', 'city-seattle']);
+
+      expect(result).toEqual([regionNode, countryNode, cityNode]);
+    });
+
+    it('should update state with fetched nodes', async () => {
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([regionNode, countryNode]);
+
+      await service.getScopeNodes(['region-us-west', 'country-usa']);
+
+      expect(service.state.nodes).toEqual({
+        'region-us-west': regionNode,
+        'country-usa': countryNode,
+      });
+    });
+
+    it('should handle empty array input', async () => {
+      const result = await service.getScopeNodes([]);
+
+      expect(result).toEqual([]);
+      expect(apiClient.fetchMultipleScopeNodes).not.toHaveBeenCalled();
+    });
+
+    it('should filter out undefined nodes', async () => {
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([]);
+
+      const result = await service.getScopeNodes(['non-existent-node']);
+
+      expect(result).toEqual([]);
+    });
+  });
+
+  describe('resolvePathToRoot with defaultPath', () => {
+    beforeEach(() => {
+      apiClient.fetchMultipleScopeNodes = jest
+        .fn()
+        .mockResolvedValue([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should use defaultPath when scope has it defined', async () => {
+      // Pre-populate scope cache
+      (service as any).updateState({
+        scopes: {
+          'scope-sea-1': scopeWithDefaultPath,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree, 'scope-sea-1');
+
+      // Should fetch all nodes in defaultPath at once
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith([
+        'region-us-west',
+        'country-usa',
+        'city-seattle',
+        'datacenter-sea-1',
+      ]);
+      expect(result.path).toEqual([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should fall back to recursive path walking when no scopeId provided', async () => {
+      // Setup nodes in cache for recursive walking
+      (service as any).updateState({
+        nodes: {
+          'datacenter-sea-1': datacenterNode,
+          'city-seattle': cityNode,
+          'country-usa': countryNode,
+          'region-us-west': regionNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree);
+
+      expect(result.path).toEqual([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should fall back when scope has no defaultPath', async () => {
+      (service as any).updateState({
+        scopes: {
+          'scope-no-path': scopeWithoutDefaultPath,
+        },
+        nodes: {
+          'datacenter-sea-1': datacenterNode,
+          'city-seattle': cityNode,
+          'country-usa': countryNode,
+          'region-us-west': regionNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree, 'scope-no-path');
+
+      expect(result.path).toEqual([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should insert path nodes into tree', async () => {
+      (service as any).updateState({
+        scopes: {
+          'scope-sea-1': scopeWithDefaultPath,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree, 'scope-sea-1');
+
+      expect(result.tree.children?.['region-us-west']).toBeDefined();
+      expect(result.tree.children?.['region-us-west']?.children?.['country-usa']).toBeDefined();
+      expect(
+        result.tree.children?.['region-us-west']?.children?.['country-usa']?.children?.['city-seattle']
+      ).toBeDefined();
+    });
+  });
+
+  describe('applyScopes with defaultPath pre-fetching', () => {
+    it('should pre-fetch all nodes from defaultPath when applying scopes', async () => {
+      apiClient.fetchMultipleScopes = jest.fn().mockResolvedValue([scopeWithDefaultPath]);
+      apiClient.fetchMultipleScopeNodes = jest
+        .fn()
+        .mockResolvedValue([regionNode, countryNode, cityNode, datacenterNode]);
+
+      await service.changeScopes(['scope-sea-1']);
+
+      // Should batch fetch all nodes in defaultPath
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith([
+        'region-us-west',
+        'country-usa',
+        'city-seattle',
+        'datacenter-sea-1',
+      ]);
+
+      // All nodes should be in cache
+      expect(service.state.nodes['region-us-west']).toEqual(regionNode);
+      expect(service.state.nodes['country-usa']).toEqual(countryNode);
+      expect(service.state.nodes['city-seattle']).toEqual(cityNode);
+      expect(service.state.nodes['datacenter-sea-1']).toEqual(datacenterNode);
+    });
+
+    it("should only pre-fetch the first scope's defaultPath", async () => {
+      const scope2: Scope = {
+        metadata: { name: 'scope-2' },
+        spec: {
+          title: 'Scope 2',
+          defaultPath: ['region-us-west', 'country-usa', 'city-portland', 'datacenter-pdx-1'],
+          filters: [],
+        },
+      };
+
+      const portlandNode: ScopeNode = {
+        metadata: { name: 'city-portland' },
+        spec: {
+          nodeType: 'container',
+          title: 'Portland',
+          parentName: 'country-usa',
+          linkType: undefined,
+          linkId: undefined,
+        },
+      };
+
+      const pdxDatacenterNode: ScopeNode = {
+        metadata: { name: 'datacenter-pdx-1' },
+        spec: {
+          nodeType: 'leaf',
+          title: 'PDX-1',
+          parentName: 'city-portland',
+          linkType: 'scope',
+          linkId: 'scope-2',
+        },
+      };
+
+      apiClient.fetchMultipleScopes = jest.fn().mockResolvedValue([scopeWithDefaultPath, scope2]);
+      apiClient.fetchMultipleScopeNodes = jest
+        .fn()
+        .mockResolvedValue([regionNode, countryNode, cityNode, datacenterNode, portlandNode, pdxDatacenterNode]);
+
+      await service.changeScopes(['scope-sea-1', 'scope-2']);
+
+      // Should only fetch the first scope's defaultPath (not the second scope's)
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith([
+        'region-us-west',
+        'country-usa',
+        'city-seattle',
+        'datacenter-sea-1',
+      ]);
+    });
+
+    it('should not fetch when scopes have no defaultPath', async () => {
+      apiClient.fetchMultipleScopes = jest.fn().mockResolvedValue([scopeWithoutDefaultPath]);
+
+      await service.changeScopes(['scope-no-path']);
+
+      expect(apiClient.fetchMultipleScopeNodes).not.toHaveBeenCalled();
+    });
+
+    it('should handle empty defaultPath array', async () => {
+      const scopeWithEmptyPath: Scope = {
+        metadata: { name: 'scope-empty' },
+        spec: {
+          title: 'Scope Empty',
+          defaultPath: [],
+          filters: [],
+        },
+      };
+
+      apiClient.fetchMultipleScopes = jest.fn().mockResolvedValue([scopeWithEmptyPath]);
+
+      await service.changeScopes(['scope-empty']);
+
+      expect(apiClient.fetchMultipleScopeNodes).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('open selector with defaultPath expansion', () => {
+    beforeEach(() => {
+      apiClient.fetchNodes = jest.fn().mockResolvedValue([regionNode]);
+      apiClient.fetchMultipleScopeNodes = jest
+        .fn()
+        .mockResolvedValue([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should expand to defaultPath when opening selector with applied scope', async () => {
+      // Apply a scope with defaultPath
+      (service as any).updateState({
+        scopes: { 'scope-sea-1': scopeWithDefaultPath },
+        appliedScopes: [{ scopeId: 'scope-sea-1', scopeNodeId: 'datacenter-sea-1' }],
+        selectedScopes: [{ scopeId: 'scope-sea-1', scopeNodeId: 'datacenter-sea-1' }],
+      });
+
+      await service.open();
+
+      // Should fetch all nodes in the path
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalled();
+
+      // Tree should be expanded to show the path
+      expect(service.state.tree.children?.['region-us-west']?.expanded).toBe(true);
+      expect(service.state.tree.children?.['region-us-west']?.children?.['country-usa']?.expanded).toBe(true);
+      expect(
+        service.state.tree.children?.['region-us-west']?.children?.['country-usa']?.children?.['city-seattle']?.expanded
+      ).toBe(true);
+    });
+
+    it('should fall back to parentNodeId when scope has no defaultPath', async () => {
+      // Pre-populate nodes for fallback behavior
+      (service as any).updateState({
+        scopes: { 'scope-no-path': scopeWithoutDefaultPath },
+        nodes: {
+          'datacenter-sea-1': datacenterNode,
+          'city-seattle': cityNode,
+          'country-usa': countryNode,
+          'region-us-west': regionNode,
+        },
+        appliedScopes: [{ scopeId: 'scope-no-path', scopeNodeId: 'datacenter-sea-1', parentNodeId: 'city-seattle' }],
+        selectedScopes: [{ scopeId: 'scope-no-path', scopeNodeId: 'datacenter-sea-1', parentNodeId: 'city-seattle' }],
+      });
+
+      await service.open();
+
+      // Should still expand, but using parentNodeId logic
+      expect(service.state.opened).toBe(true);
+    });
+
+    it('should handle opening selector when scope is not yet loaded', async () => {
+      (service as any).updateState({
+        appliedScopes: [{ scopeId: 'scope-sea-1' }],
+        selectedScopes: [{ scopeId: 'scope-sea-1' }],
+      });
+
+      await service.open();
+
+      // Should not crash, just open with root nodes
+      expect(service.state.opened).toBe(true);
+    });
+  });
+
+  describe('performance improvements', () => {
+    it('should make only 1 API call for deep hierarchy with defaultPath', async () => {
+      (service as any).updateState({
+        scopes: { 'scope-sea-1': scopeWithDefaultPath },
+      });
+
+      apiClient.fetchMultipleScopeNodes = jest
+        .fn()
+        .mockResolvedValue([regionNode, countryNode, cityNode, datacenterNode]);
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      await service.resolvePathToRoot('datacenter-sea-1', tree, 'scope-sea-1');
+
+      // Should make exactly 1 API call
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledTimes(1);
+    });
+
+    it('should make N API calls for deep hierarchy without defaultPath (old behavior)', async () => {
+      // This test documents the old recursive behavior for comparison
+      apiClient.fetchScopeNode = jest.fn().mockImplementation((id: string) => {
+        const nodeMap: Record<string, ScopeNode> = {
+          'datacenter-sea-1': datacenterNode,
+          'city-seattle': cityNode,
+          'country-usa': countryNode,
+          'region-us-west': regionNode,
+        };
+        return Promise.resolve(nodeMap[id]);
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      await service.resolvePathToRoot('datacenter-sea-1', tree);
+
+      // Would make 4 sequential calls in the old implementation
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('datacenter-sea-1');
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('city-seattle');
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('country-usa');
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('region-us-west');
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledTimes(4);
+    });
+  });
+
+  describe('edge cases and error handling', () => {
+    it('should handle defaultPath with missing nodes gracefully', async () => {
+      (service as any).updateState({
+        scopes: { 'scope-sea-1': scopeWithDefaultPath },
+      });
+
+      // API returns fewer nodes than requested
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([regionNode, countryNode]);
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree, 'scope-sea-1');
+
+      // Should handle partial path gracefully
+      expect(result.path).toEqual([regionNode, countryNode]);
+    });
+
+    it('should handle API errors during batch fetch', async () => {
+      (service as any).updateState({
+        scopes: { 'scope-sea-1': scopeWithDefaultPath },
+      });
+
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([]);
+
+      await service.changeScopes(['scope-sea-1']);
+
+      // Should not crash, state should be consistent
+      expect(service.state.appliedScopes).toEqual([{ scopeId: 'scope-sea-1' }]);
+    });
+
+    it('should deduplicate node IDs in defaultPath', async () => {
+      const scopeWithDuplicates: Scope = {
+        metadata: { name: 'scope-dupe' },
+        spec: {
+          title: 'Scope with Duplicates',
+          defaultPath: ['region-us-west', 'country-usa', 'region-us-west', 'country-usa'],
+          filters: [],
+        },
+      };
+
+      apiClient.fetchMultipleScopes = jest.fn().mockResolvedValue([scopeWithDuplicates]);
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([regionNode, countryNode]);
+
+      await service.changeScopes(['scope-dupe']);
+
+      // Should only fetch unique nodes
+      const calledWith = apiClient.fetchMultipleScopeNodes.mock.calls[0][0];
+      const uniqueNodes = [...new Set(calledWith)];
+      expect(calledWith.length).toBe(uniqueNodes.length);
+    });
+
+    it('should handle defaultPath with only root node', async () => {
+      const scopeWithRootOnly: Scope = {
+        metadata: { name: 'scope-root' },
+        spec: {
+          title: 'Scope Root Only',
+          defaultPath: ['region-us-west'],
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'scope-root': scopeWithRootOnly },
+      });
+
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([regionNode]);
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('region-us-west', tree, 'scope-root');
+
+      expect(result.path).toEqual([regionNode]);
+    });
+  });
+
+  describe('backwards compatibility', () => {
+    it('should work with existing code that does not provide scopeId to resolvePathToRoot', async () => {
+      (service as any).updateState({
+        nodes: {
+          'datacenter-sea-1': datacenterNode,
+          'city-seattle': cityNode,
+          'country-usa': countryNode,
+          'region-us-west': regionNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('datacenter-sea-1', tree);
+
+      expect(result.path).toEqual([regionNode, countryNode, cityNode, datacenterNode]);
+    });
+
+    it('should not break when scope metadata is loaded after applying', async () => {
+      // This simulates the async nature of scope loading
+      apiClient.fetchMultipleScopes = jest.fn().mockImplementation(async () => {
+        // Simulate delay
+        await new Promise((resolve) => setTimeout(resolve, 10));
+        return [scopeWithDefaultPath];
+      });
+
+      await service.changeScopes(['scope-sea-1']);
+
+      // Scope should eventually be in state
+      expect(service.state.scopes['scope-sea-1']).toEqual(scopeWithDefaultPath);
+    });
+  });
+
+  // Tests for path helper methods
+  describe('getPathForScope (new helper method)', () => {
+    it('should prefer defaultPath from scope metadata', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child'],
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+      });
+
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([parentNode, childNode]);
+
+      // This tests the new getPathForScope method that should be created
+      // For now, this is testing the expected behavior through resolvePathToRoot
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('child', tree, 'test-scope');
+
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith(['parent', 'child']);
+      expect(result.path).toEqual([parentNode, childNode]);
+    });
+
+    it('should fall back to scopeNodeId when no defaultPath', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        nodes: {
+          parent: parentNode,
+          child: childNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('child', tree, 'test-scope');
+
+      expect(result.path).toEqual([parentNode, childNode]);
+    });
+
+    it('should return empty array when both scopeId and scopeNodeId are undefined', async () => {
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('', tree);
+
+      expect(result.path).toEqual([]);
+    });
+
+    it('should handle scope not being in cache yet', async () => {
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      (service as any).updateState({
+        nodes: {
+          parent: parentNode,
+          child: childNode,
+        },
+      });
+
+      // Scope not in cache, but scopeNodeId is provided
+      const result = await service.resolvePathToRoot('child', tree, 'unknown-scope');
+
+      // Should fall back to node-based path
+      expect(result.path).toEqual([parentNode, childNode]);
+    });
+  });
+
+  describe('getNodePath - optimized implementation', () => {
+    it('should build path from cached nodes without API calls', async () => {
+      (service as any).updateState({
+        nodes: {
+          parent: parentNode,
+          child: childNode,
+          grandchild: grandchildNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('grandchild', tree);
+
+      // Should not make any API calls since all nodes are cached
+      expect(apiClient.fetchScopeNode).not.toHaveBeenCalled();
+      expect(result.path).toEqual([parentNode, childNode, grandchildNode]);
+    });
+
+    it('should fetch missing nodes in the path', async () => {
+      // Only grandchild is cached
+      (service as any).updateState({
+        nodes: {
+          grandchild: grandchildNode,
+        },
+      });
+
+      apiClient.fetchScopeNode = jest.fn().mockImplementation((id: string) => {
+        if (id === 'child') {
+          return Promise.resolve(childNode);
+        }
+        if (id === 'parent') {
+          return Promise.resolve(parentNode);
+        }
+        return Promise.resolve(undefined);
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      await service.resolvePathToRoot('grandchild', tree);
+
+      // Should fetch missing parent nodes
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('child');
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('parent');
+    });
+
+    it('should handle circular references gracefully', async () => {
+      const circularNode1: ScopeNode = {
+        metadata: { name: 'node1' },
+        spec: {
+          nodeType: 'container',
+          title: 'Node 1',
+          parentName: 'node2',
+        },
+      };
+
+      const circularNode2: ScopeNode = {
+        metadata: { name: 'node2' },
+        spec: {
+          nodeType: 'container',
+          title: 'Node 2',
+          parentName: 'node1',
+        },
+      };
+
+      (service as any).updateState({
+        nodes: {
+          node1: circularNode1,
+          node2: circularNode2,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      // This should not hang or crash
+      // Implementation should detect circular references and stop recursion
+      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
+      const result = await service.resolvePathToRoot('node1', tree);
+
+      expect(result).toBeDefined();
+      // When circular reference is detected, it returns partial path (up to the circular point)
+      expect(result.path.length).toBeGreaterThan(0);
+      expect(consoleErrorSpy).toHaveBeenCalledWith('Circular reference detected in node path', expect.any(String));
+      consoleErrorSpy.mockRestore();
+    });
+
+    it('should stop at root node (empty parentName)', async () => {
+      (service as any).updateState({
+        nodes: {
+          parent: parentNode,
+          child: childNode,
+        },
+      });
+
+      const tree = {
+        expanded: false,
+        scopeNodeId: '',
+        query: '',
+        children: {},
+      };
+
+      const result = await service.resolvePathToRoot('child', tree);
+
+      expect(result.path).toEqual([parentNode, childNode]);
+      expect(result.path[0].spec.parentName).toBe('');
+    });
+  });
+
+  describe('expandToSelectedScope (new helper method)', () => {
+    beforeEach(() => {
+      apiClient.fetchNodes = jest.fn().mockResolvedValue([parentNode]);
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([parentNode, childNode]);
+    });
+
+    it('should expand tree to show selected scope path', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child'],
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        selectedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+        appliedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+      });
+
+      await service.open();
+
+      // Tree should be expanded to show the path
+      expect(service.state.tree.children?.['parent']?.expanded).toBe(true);
+      expect(service.state.tree.children?.['parent']?.children?.['child']).toBeDefined();
+    });
+
+    it('should not expand when no scopes are selected', async () => {
+      (service as any).updateState({
+        selectedScopes: [],
+        appliedScopes: [],
+      });
+
+      await service.open();
+
+      // Root should have children loaded but not expanded beyond that
+      expect(service.state.tree.children).toBeDefined();
+    });
+
+    it('should load children of the last node in the path', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child'],
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        selectedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child', parentNodeId: 'parent' }],
+        appliedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child', parentNodeId: 'parent' }],
+      });
+
+      // Mock API to return path nodes
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([parentNode, childNode]);
+
+      // Mock fetchNodes to return children of the last node
+      apiClient.fetchNodes = jest.fn().mockImplementation((options) => {
+        if (options.parent === '') {
+          return Promise.resolve([parentNode]);
+        } else if (options.parent === 'parent') {
+          return Promise.resolve([childNode]);
+        }
+        return Promise.resolve([]);
+      });
+
+      await service.open();
+
+      // Should have loaded root children (called once during tree initialization)
+      expect(apiClient.fetchNodes).toHaveBeenCalled();
+      // Verify the path nodes were fetched (parent already in cache from fetchNodes, so only child is fetched)
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith(['child']);
+    });
+
+    it('should handle errors gracefully when expanding', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child'],
+          filters: [],
+        },
+      };
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        selectedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+        appliedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+      });
+
+      // Mock API to fail
+      const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation();
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockRejectedValue(new Error('API Error'));
+
+      // Should not crash
+      await expect(service.open()).resolves.not.toThrow();
+      expect(service.state.opened).toBe(true);
+      consoleErrorSpy.mockRestore();
+    });
+  });
+
+  describe('integration - full path resolution flow', () => {
+    it('should resolve path from defaultPath, insert into tree, and expand', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child', 'grandchild'],
+          filters: [],
+        },
+      };
+
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([parentNode, childNode, grandchildNode]);
+      apiClient.fetchNodes = jest.fn().mockResolvedValue([parentNode]);
+
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        selectedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'grandchild' }],
+        appliedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'grandchild' }],
+      });
+
+      await service.open();
+
+      // Path should be resolved (parent already in cache from fetchNodes, so only child and grandchild are fetched)
+      expect(apiClient.fetchMultipleScopeNodes).toHaveBeenCalledWith(['child', 'grandchild']);
+
+      // Nodes should be in cache
+      expect(service.state.nodes['parent']).toEqual(parentNode);
+      expect(service.state.nodes['child']).toEqual(childNode);
+      expect(service.state.nodes['grandchild']).toEqual(grandchildNode);
+
+      // Tree should be expanded
+      expect(service.state.tree.children?.['parent']?.expanded).toBe(true);
+      expect(service.state.tree.children?.['parent']?.children?.['child']?.expanded).toBe(true);
+    });
+
+    it('should use cached nodes and avoid unnecessary API calls', async () => {
+      const scope: Scope = {
+        metadata: { name: 'test-scope' },
+        spec: {
+          title: 'Test Scope',
+          defaultPath: ['parent', 'child'],
+          filters: [],
+        },
+      };
+
+      // Pre-populate cache
+      (service as any).updateState({
+        scopes: { 'test-scope': scope },
+        nodes: {
+          parent: parentNode,
+          child: childNode,
+        },
+        selectedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+        appliedScopes: [{ scopeId: 'test-scope', scopeNodeId: 'child' }],
+      });
+
+      apiClient.fetchNodes = jest.fn().mockResolvedValue([parentNode]);
+      apiClient.fetchMultipleScopeNodes = jest.fn().mockResolvedValue([]);
+
+      await service.open();
+
+      // Should not fetch nodes that are already cached
+      expect(apiClient.fetchMultipleScopeNodes).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('getScopeNode - caching behavior', () => {
+    it('should return cached node without API call', async () => {
+      (service as any).updateState({
+        nodes: {
+          'test-node': childNode,
+        },
+      });
+
+      const result = await service.getScopeNode('test-node');
+
+      expect(result).toEqual(childNode);
+      expect(apiClient.fetchScopeNode).not.toHaveBeenCalled();
+    });
+
+    it('should fetch and cache node when not in cache', async () => {
+      apiClient.fetchScopeNode = jest.fn().mockResolvedValue(childNode);
+
+      const result = await service.getScopeNode('test-node');
+
+      expect(apiClient.fetchScopeNode).toHaveBeenCalledWith('test-node');
+      expect(result).toEqual(childNode);
+      // Node is cached using its metadata.name, not the requested ID
+      expect(service.state.nodes['child']).toEqual(childNode);
+    });
+
+    it('should handle API errors gracefully', async () => {
+      apiClient.fetchScopeNode = jest.fn().mockResolvedValue(undefined);
+
+      const result = await service.getScopeNode('non-existent');
+
+      expect(result).toBeUndefined();
+      expect(service.state.nodes['non-existent']).toBeUndefined();
     });
   });
 });
